feat: gate reconciliation, MPC audit remediation, OPEN problem resolution
Summary (cumulative session):
- T1-T7 gate reconciliation: phase1/phase2 GREEN, phase3 9/12 + 3 CI-deferred
- legacy-fold poison-pill eliminated
- Shamir threshold bound fixed: (n-1)/2 → floor(n/2)+1
- F67 decrypt-share wire-v2: ct_hash binding + party_id validation
- compressor doctest fences fixed, spec-tests license added
- latticefold E0317 regression fixed (compressor_glue.rs)

execution-wave-1 (5/7 waves complete):
- C5 formation proof: PoP-based pk_agg = Σ pk_i (9 tests)
- C7 correctness: Schwartz-Zippel Lagrange recombination in Noir (14 tests)
- A1 accumulator: versioned codec, fail-closed removed (21 tests)
- G4 PK binding: Merkle-path in-circuit (18 Noir tests)
- G3 plaintext binding: aggregate_decrypt_raw_result_poly wired

mpc-audit-2026-06-04 remediation (20/20 P0-P2 findings):
- P0 (HIGH): empty bootstrap proof reject, zero-step IVC block, H2 commit verify,
  IvcBinding fields in statement hash (Sol+Rust+Noir lockstep), Nova session binding
- P1 (MEDIUM): Nova no-ops/idle/counters/prove fixes, wire size bounds,
  encode_fields Result, backend session binding
- P2 (MEDIUM): NIZK Fiat-Shamir d_commitment binding, bootstrap round index,
  BFV sigma encoding, Solidity cross-session adversarial tests (156/156 forge)

meta-plan coverage:
- meta-plan-all-deferred updated with full OPEN problem coverage (Phase H)
- 3 sub-plans created: c5-formation-proof, c7-correctness, a1-accumulator-transcript
- execution-wave-1 plan (23/23 tasks done)

OPEN problems resolved: C5 ✅, C7 ✅, A1 ✅
OPEN problems status unchanged: P1, P2, P4 (fail-closed)
Test suites: all green (Sol 156/156, Noir 18/18, Rust 62+75+20+9=166+)

+++ crates/pvthfhe-compressor/src/nova/monomial_range.rs
@@ -42,7 +42,7 @@ pub fn monomial_range_check_bp<CS: ConstraintSystem<NovaScalar>>(
 
     let bits: Vec<AllocatedNum<NovaScalar>> = (0..num_bits)
         .map(|idx| {
-            let bit_val = NovaScalar::from(((value >> idx) & 1) as u64);
+            let bit_val = NovaScalar::from((value >> idx) & 1);
             AllocatedNum::alloc(cs.namespace(|| format!("{tag}_m_bit_{idx}")), || {
                 Ok(bit_val)
             })
@@ -50,7 +50,7 @@ pub fn monomial_range_check_bp<CS: ConstraintSystem<NovaScalar>>(
         .collect::<Result<_, _>>()?;
 
     for idx in 0..num_bits {
-        let bit_val = NovaScalar::from(((value >> idx) & 1) as u64);
+        let bit_val = NovaScalar::from((value >> idx) & 1);
 
         let bit_minus_one_val = if bit_val == NovaScalar::ONE {
             NovaScalar::ZERO
@@ -106,7 +106,7 @@ pub fn monomial_range_check_bp<CS: ConstraintSystem<NovaScalar>>(
     );
 
     if bound > 1 && !bound.is_power_of_two() {
-        let bound_msb = (usize::BITS as u32 - bound.leading_zeros() - 1) as usize;
+        let bound_msb = (usize::BITS - bound.leading_zeros() - 1) as usize;
         if num_bits > bound_msb {
             for idx in bound_msb..num_bits {
                 let zero_val =
@@ -134,7 +134,7 @@ fn monomial_bit_count(bound: u64) -> usize {
     if bound <= 1 {
         return 1;
     }
-    (usize::BITS as u32 - (bound - 1).leading_zeros()) as usize
+    (usize::BITS - (bound - 1).leading_zeros()) as usize
 }
 
 /// Precomputed monomial table polynomial for embedding-based range proofs.
